Add image names to docker-compose file
This allows the images to have a name and tag when built. Having both the `build`
and `image` argument in the compose file means that the image gets named
correctly and can then be viewed with `docker images`.

https://docs.docker.com/compose/compose-file/#build

When using `kompose`, it also means the image name is added to the k8s
manifest deployment files.