Add k8s support
Using `kompose`, I created a bunch of manifest files from the
docker-compose.yml file.

I did previously try adding the `command` parameter in a few deployment files
based off what is within the `start-docker.sh` script, but this caused
problems booting the containers, and I don't really know enough about
what was happening.

Kompose also didn't create a PersistentVolume to go along with the
PersistentVolumeClaims, so I had to also add this in so it played well
with minikube. If this is booted into an external infrastructure then it
may well need to be extracted.