Add selenium-webdriver and webdrivers gems
These gems are added as they're not included by default with govuk_test
anymore. Webdrivers acts as a replacement for chromedriver-helper.

The code for looking up chromedriver path should not be needed anymore
as this path can be set as an environment variable in
webdrivers: https://github.com/titusfortner/webdrivers#download-location

This also removes some old config from capybara/webkit