Configure Chrome using Selenium Options object
This resolves a problem that the key to provide options to Chrome has
changed from "chromeOptions" to "goog:chromeOptions". To try reduce our
chances of hitting a similar problem again I've switched to using the
Selenium object to manage chrome options.

I'm not too sure if this will introduce any problems with the
allowInsecureCerts as that isn't a config that is used as part of the
e2e tests.

I've removed a number of options that I couldn't find a reason they were
added. I figure if they're needed we can re-add them and explain why
they are needed in the commit message so that is documented.