Set ASSET_HOST environment variable for government frontend.
This is to allows assets to continue to work when they are moved from
the assets host to the www one. Rather than add a bunch of nginx config
I decided to follow the approach from govuk-docker and instead set the
asset hostname to match the applications so they no proxying is
required.