Select HMRC when creating a new contact
The current finder for HMRC contacts does not actually filter by
organisation and shows contacts for all organisations. This means the
e2e test for this will pass regardless of the organisation selected.

In https://github.com/alphagov/contacts-admin/pull/827, we are filtering
to only show HMRC contacts, so need to select that organisation when
creating the contact.