Replace Docker::Compose with shell backticks
Using Docker::Compose.new.ps has been causing us intermittent errors in
recent weeks (since nginx-proxy was last updated) which has been causing
exceptions such as:

```
13:16:58  Docker::Compose::Error: 'docker ps' failed with status 'pid 15874 exit 0': Cannot parse output: '(a7a18ebdc64e1dd0a40f68bd62aa4b1126c68fb9e945a6857ede159e377da519) (jwilder/nginx-proxy:latest) (18.2kB (virtual 158MB)) (Up 27 seconds) (lishing-e2e-tests_test-against_2_nginx-proxy_1) (com.docker.compose.project=lishing-e2e-tests_test-against_2,org.opencontainers.image.created=2021-04-12T00:31:10.992Z,org.opencontainers.image.licenses=MIT,org.opencontainers.image.url=https://github.com/nginx-proxy/nginx-proxy,org.opencontainers.image.version=latest,com.docker.compose.config-hash=5517769695a167132a40ede6634a33e23c406504e4697fc1ff2f8924a11aba09,com.docker.compose.oneoff=False,com.docker.compose.version=1.22.0,org.opencontainers.image.source=https://github.com/nginx-proxy/nginx-proxy,org.opencontainers.image.title=nginx-proxy,maintainer=Jason Wilder mail@jasonwilder.com,org.opencontainers.image.description=Automated nginx proxy for Docker containers using docker-gen,org.opencontainers.image.revision=6613e272eb41c926605bc0411b7af9f70c2ebeee,com.docker.compose.container-number=1,com.docker.compose.service=nginx-'
```

We have determined that this problem is risked whenever a docker ps
command returns a string longer than 1024 characters (as the recent
update to nginx-proxy now causes) due to a supporting gem, backticks,
which seems to have a race condition, which I have raised an issue for:
https://github.com/xeger/backticks/issues/17

To resolve this I have changed from using the Docker Compose gem to
making the calls to docker-compose with Ruby backticks. This uses a call
to `docker-compose ps --services` to list all of the services in the
docker compose file (it seems to list them all regardless of whether
they are running) and then using `docker-compose ps -q <service-name>`
to get the container id of that particular service.